+++ src/vīcĭus.vcxproj.filters
@@ -93,6 +93,9 @@
     <ClCompile Include="markdown.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="InstanceConfig.Postpone.cpp">
+      <Filter>Source Files\Models</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="resource.h">
